feat(user-roles-permissions): initial implementation of roles and permission

+++ server/my-family-ws/src/main/java/com/bytepushers/family/dao/UserJdbcDAO.java
@@ -70,7 +70,7 @@ public class UserJdbcDAO implements UserDAO {
             ResultSet rs = stmt.getGeneratedKeys();
             if (rs.next()) {
                 createdUser = new User(someUser.getEmail(), someUser.getPassword(), someUser.getUsername(), someUser.getEnabled()); // TODO: Create a findUserByID, pass in PK
-                createdUser.setId(rs.getInt(1));  // Get the generated ID and set it
+                createdUser.setId(rs.getLong(1));  // Get the generated ID and set it
                 logger.info("User created with id: {}", createdUser.getId());  // Happy path log
             }
         } catch (SQLException e) {
@@ -94,7 +94,7 @@ public class UserJdbcDAO implements UserDAO {
     }
 
     @Override
-    public User findUserById(Integer id) {
+    public User findUserById(Long id) {
         String sql = "SELECT * FROM users WHERE id = ?";
 
         try (Connection conn = getConnection();
@@ -176,7 +176,7 @@ public class UserJdbcDAO implements UserDAO {
     }
 
     @Override
-    public boolean deleteUser(Integer id) {
+    public boolean deleteUser(Long id) {
         String sql = "DELETE FROM users WHERE id = ?";
 
         try (Connection conn = getConnection();
@@ -200,7 +200,7 @@ public class UserJdbcDAO implements UserDAO {
     // Helper method to map a ResultSet row to a User object
     private User mapResultSetToUser(ResultSet rs) throws SQLException {
         User user = new User();
-        user.setId(rs.getInt("id"));
+        user.setId(rs.getLong("id"));
         user.setEmail(rs.getString("email"));
         user.setPassword(rs.getString("password"));
         user.setUsername(rs.getString("username"));
